separated groovy ast class from registration into different modules
