+++ webtau-core-groovy/src/main/groovy/com/twosigma/webtau/data/expectation/ShouldNot.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.twosigma.webtau.groovy.expectation
+package com.twosigma.webtau.data.expectation
 
 class ShouldNot {
     private Object actual
