+++ webtau-core-groovy/src/main/resources/META-INF/services/org.codehaus.groovy.runtime.ExtensionModule
@@ -16,4 +16,4 @@
 
 moduleName=webtau-core-groovy
 moduleVersion=1.0
-extensionClasses=com.twosigma.webtau.data.table.TableDataExtension,com.twosigma.webtau.data.time.TestTimeExtension
+extensionClasses=com.twosigma.webtau.data.expectation.ExpectationExtension,com.twosigma.webtau.data.table.TableDataExtension,com.twosigma.webtau.data.time.TestTimeExtension
