--- webtau-groovy-ast/src/main/resources/META-INF/services/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,19 +1,0 @@
-#
-# Copyright 2018 TWO SIGMA OPEN SOURCE, LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-moduleName=webtau-groovy-ast
-moduleVersion=1.0
-extensionClasses=com.twosigma.webtau.groovy.expectation.ExpectationExtension
